ATR-767: migrated override and pxoverride tests
