+++ src/Acuminator/Acuminator.Tests/Tests/StaticAnalysis/NameConventionEventsInGraphsAndGraphExtensions/Sources/PXOverride.cs
@@ -5,31 +5,15 @@ using System.Text;
 using System.Threading.Tasks;
 using PX.Data;
 
-namespace PX.Objects
+namespace Acuminator.Analyzers.StaticAnalysis.NameConventionEventsInGraphsAndGraphExtensions.Sources
 {
-	public class SOInvoiceEntryExt : PXGraphExtension<SOInvoiceEntry>
+	// Acuminator disable once PX1016 ExtensionDoesNotDeclareIsActiveMethod extension should be constantly active
+	public class SOInvoiceEntryExt : PXGraphExtension<SOInvoiceCustomEntry>
 	{
 		[PXOverride]
-		public virtual void SOInvoice_RowSelected(PXCache cache, PXRowSelectedEventArgs e)
+		public virtual void SOInvoice_RowSelected(PXCache cache, PXRowSelectedEventArgs e)		// Should not show PX1041
 		{
 			
 		}
-	}
-
-	public class SOInvoiceEntry : SOInvoiceEntryBase
-	{
-		protected virtual void SOInvoice_RowSelected(PXCache cache, PXRowSelectedEventArgs e)
-		{
-			
-		}
-	}
-
-	public class SOInvoice : IBqlTable
-	{
-		#region RefNbr
-		[PXDBString(8, IsKey = true, InputMask = "")]
-		public string RefNbr { get; set; }
-		public abstract class refNbr : IBqlField { }
-		#endregion	
 	}
 }
